pyproject.toml
[project]
name = "portfoyar"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "saraey", email = "saraeygh@gmail.com"},
]
dependencies = [
    "Django==5.0.1",
    "celery==5.3.6",
    "beautifulsoup4==4.12.3",
    "django-cors-headers==4.3.1",
    "django-extensions==3.2.3",
    "django-jalali==6.0.1",
    "djangorestframework==3.14.0",
    "drf-excel==2.4.0",
    "gunicorn==21.2.0",
    "jdatetime==4.1.1",
    "locust==2.23.1",
    "matplotlib==3.8.2",
    "pandas==2.2.0",
    "persiantools==3.0.1",
    "psycopg2-binary==2.9.9",
    "pymongo==4.6.1",
    "redis==5.0.1",
    "requests==2.31.0",
    "tqdm==2.2.3",
    "drf-api-logger>=1.1.16",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
